Rework osquery configuration to be derived from an array
OSQuery uses a JSON configuration file to store it's configuration.
Currently, this is a hard-coded set of configuration that is injected in
it's entirety. However, this does not allow for easy reuse of this role
across different machines, each of which may have different profiles to
insert this content.

This commit replaces it with an array generated from Ansible
configuration yaml. Yaml exists as basically a superset of JSON, and
Ansible filters allow easy conversion between the two syntaxes. This
allows almost infinite flexibility, while maintaining the expected and
understood configuration tree.